moved the api endpoints to thier own file (routes) for seperations of concerns and for unit testing purposes
